+++ lib/queries/source/autocomplete.rb
@@ -15,7 +15,7 @@ module Queries
 
       # @return [Scope]
       def base_query
-        ::Source.select('sources.*')
+        ::Source #.select('sources.*')
       end
 
       # @return [ActiveRecord::Relation]
@@ -185,7 +185,15 @@ module Queries
 
         updated_queries = []
         queries.each_with_index do |q ,i|
-          a = q.joins(:project_sources).where(member_of_project_id.to_sql) if project_id && limit_to_project
+          if project_id && limit_to_project
+            a = q.joins(:project_sources).where(member_of_project_id.to_sql)
+          else
+            a = q.left_outer_joins(:citations)
+                 .select('sources.*, COUNT(citations.id) AS use_count, MAX(citations.project_id) AS in_project_id')
+                 .where('citations.project_id = ? OR citations.project_id IS NULL', Current.project_id)
+                 .group('sources.id')
+                 .order('in_project_id, use_count DESC')
+          end
           a ||= q
           updated_queries[i] = a
         end